Authenticator/requirements.txt
python-dotenv
Flask-RESTful
PyJWT
Flask-Bcrypt
marshmallow
Flask-SQLAlchemy
flask-cors